Use u32 insead of u16 for unique ids
Also use AtomicU32 instead of the arc-mutex-int combination

+++ applications/lobsters/pelton/endpoints/pelton/comment.rs
@@ -11,8 +11,8 @@ pub(crate) async fn handle<F>(
     story: StoryId,
     parent: Option<CommentId>,
     priming: bool,
-    comment_uid: u16,
-    vote_uid: u16,
+    comment_uid: u32,
+    vote_uid: u32,
 ) -> Result<(my::Conn, bool), my::error::Error>
 where
     F: 'static + Future<Output = Result<my::Conn, my::error::Error>> + Send,
